requirements.txt
# Asosiy kutubxonalar
aiogram>=3.1.0,<3.3.0
python-dotenv>=0.19.0

# Ma'lumotlar bazasi
aiosqlite>=0.17.0

# Audio processing
pydub>=0.25.0
ffmpeg-python>=0.2.0

# Logging va monitoring
loguru>=0.6.0

# Utilities
python-multipart>=0.0.5
typing-extensions>=4.0.0

# Development dependencies (ixtiyoriy)
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0

# Optional: Web interface uchun (kelajakda)
fastapi>=0.100.0
uvicorn>=0.20.0
